Use wlp kernel for wlp tests
