+++ src/integTest/resources/kernel-install-feature-test/build_wlp.gradle
@@ -1,0 +1,24 @@
+apply plugin: 'liberty'
+
+buildscript {
+    repositories{
+      mavenLocal()
+      mavenCentral()
+      maven {
+            name = 'Sonatype Nexus Snapshots'
+            url = 'https://oss.sonatype.org/content/repositories/snapshots/'
+      }
+    }
+    dependencies {
+        classpath group: 'net.wasdev.wlp.gradle.plugins', name: 'liberty-gradle-plugin', version: lgpVersion
+    }
+}
+
+repositories {
+      mavenLocal()
+      mavenCentral()
+}
+
+dependencies {
+    libertyRuntime group: 'com.ibm.websphere.appserver.runtime', name: 'wlp-kernel', version: '18.0.0.2'
+}
